+++ R/facet-wrap.r
@@ -147,29 +147,29 @@ FacetWrap <- ggproto("FacetWrap", Facet,
       return(layout_null())
     }
 
-    base <- unrowname(
-      combine_vars(data, params$plot_env, vars, drop = params$drop)
-    )
+    base <- combine_vars(data, params$plot_env, vars, drop = params$drop)
 
     id <- id(base, drop = TRUE)
     n <- attr(id, "n")
 
     dims <- wrap_dims(n, params$nrow, params$ncol)
-    layout <- new_data_frame(list(PANEL = factor(id, levels = seq_len(n))))
-
-    if (params$as.table) {
-      layout$ROW <- as.integer((id - 1L) %/% dims[2] + 1L)
-    } else {
-      layout$ROW <- as.integer(dims[1] - (id - 1L) %/% dims[2])
-    }
-    layout$COL <- as.integer((id - 1L) %% dims[2] + 1L)
+    layout <- data_frame0(
+      PANEL = factor(id, levels = seq_len(n)),
+      ROW = if (params$as.table) {
+        as.integer((id - 1L) %/% dims[2] + 1L)
+      } else {
+        as.integer(dims[1] - (id - 1L) %/% dims[2])
+      },
+      COL = as.integer((id - 1L) %% dims[2] + 1L),
+      .size = length(id)
+    )
 
     # For vertical direction, flip row and col
     if (identical(params$dir, "v")) {
       layout[c("ROW", "COL")] <- layout[c("COL", "ROW")]
     }
 
-    panels <- cbind(layout, unrowname(base))
+    panels <- vec_cbind(layout, base)
     panels <- panels[order(panels$PANEL), , drop = FALSE]
     rownames(panels) <- NULL
 
@@ -181,7 +181,7 @@ FacetWrap <- ggproto("FacetWrap", Facet,
   },
   map_data = function(data, layout, params) {
     if (empty(data)) {
-      return(cbind(data, PANEL = integer(0)))
+      return(vec_cbind(data %|W|% NULL, PANEL = integer(0)))
     }
 
     vars <- params$facets
@@ -193,19 +193,21 @@ FacetWrap <- ggproto("FacetWrap", Facet,
 
     facet_vals <- eval_facets(vars, data, params$.possible_columns)
     facet_vals[] <- lapply(facet_vals[], as.factor)
+    layout[] <- lapply(layout[], as.factor)
 
     missing_facets <- setdiff(names(vars), names(facet_vals))
     if (length(missing_facets) > 0) {
 
-      to_add <- unique(layout[missing_facets])
+      to_add <- unique0(layout[missing_facets])
 
       data_rep <- rep.int(1:nrow(data), nrow(to_add))
       facet_rep <- rep(1:nrow(to_add), each = nrow(data))
 
-      data <- unrowname(data[data_rep, , drop = FALSE])
-      facet_vals <- unrowname(cbind(
+      data <- data[data_rep, , drop = FALSE]
+      facet_vals <- vec_cbind(
         facet_vals[data_rep, ,  drop = FALSE],
-        to_add[facet_rep, , drop = FALSE]))
+        to_add[facet_rep, , drop = FALSE]
+      )
     }
 
     keys <- join_keys(facet_vals, layout, by = names(vars))
@@ -243,7 +245,7 @@ FacetWrap <- ggproto("FacetWrap", Facet,
 
     if (length(params$facets) == 0) {
       # Add a dummy label
-      labels_df <- new_data_frame(list("(all)" = "(all)"), n = 1)
+      labels_df <- data_frame0("(all)" = "(all)", .size = 1)
     } else {
       labels_df <- layout[names(params$facets)]
     }
@@ -322,10 +324,12 @@ FacetWrap <- ggproto("FacetWrap", Facet,
       empty_bottom <- apply(empties, 2, function(x) c(diff(x) == 1, FALSE))
       if (any(empty_bottom)) {
         pos <- which(empty_bottom)
-        panel_loc <- data_frame(ROW = row_ind[pos], COL = col_ind[pos])
-        # Substitute with vctrs::vec_match(panel_loc, layout[, c("ROW", "COL")])
-        # Once we switch to vctrs wholesale
-        panels <- merge(panel_loc, cbind(layout, .index = seq_len(nrow(layout))))$.index
+        panel_loc <- data_frame0(
+          ROW = row_ind[pos],
+          COL = col_ind[pos],
+          .size = length(pos)
+        )
+        panels <- vec_match(panel_loc, layout[, c("ROW", "COL")])
         x_axes <- axes$x$bottom[layout$SCALE_X[panels]]
         if (params$strip.position == "bottom" &&
             !inside &&
@@ -339,8 +343,12 @@ FacetWrap <- ggproto("FacetWrap", Facet,
       empty_top <- apply(empties, 2, function(x) c(FALSE, diff(x) == -1))
       if (any(empty_top)) {
         pos <- which(empty_top)
-        panel_loc <- data_frame(ROW = row_ind[pos], COL = col_ind[pos])
-        panels <- merge(panel_loc, cbind(layout, .index = seq_len(nrow(layout))))$.index
+        panel_loc <- data_frame0(
+          ROW = row_ind[pos],
+          COL = col_ind[pos],
+          .size = length(pos)
+        )
+        panels <- vec_match(panel_loc, layout[, c("ROW", "COL")])
         x_axes <- axes$x$top[layout$SCALE_X[panels]]
         if (params$strip.position == "top" &&
             !inside &&
@@ -354,8 +362,12 @@ FacetWrap <- ggproto("FacetWrap", Facet,
       empty_right <- t(apply(empties, 1, function(x) c(diff(x) == 1, FALSE)))
       if (any(empty_right)) {
         pos <- which(empty_right)
-        panel_loc <- data_frame(ROW = row_ind[pos], COL = col_ind[pos])
-        panels <- merge(panel_loc, cbind(layout, .index = seq_len(nrow(layout))))$.index
+        panel_loc <- data_frame0(
+          ROW = row_ind[pos],
+          COL = col_ind[pos],
+          .size = length(pos)
+        )
+        panels <- vec_match(panel_loc, layout[, c("ROW", "COL")])
         y_axes <- axes$y$right[layout$SCALE_Y[panels]]
         if (params$strip.position == "right" &&
             !inside &&
@@ -369,8 +381,12 @@ FacetWrap <- ggproto("FacetWrap", Facet,
       empty_left <- t(apply(empties, 1, function(x) c(FALSE, diff(x) == -1)))
       if (any(empty_left)) {
         pos <- which(empty_left)
-        panel_loc <- data_frame(ROW = row_ind[pos], COL = col_ind[pos])
-        panels <- merge(panel_loc, cbind(layout, .index = seq_len(nrow(layout))))$.index
+        panel_loc <- data_frame0(
+          ROW = row_ind[pos],
+          COL = col_ind[pos],
+          .size = length(pos)
+        )
+        panels <- vec_match(panel_loc, layout[, c("ROW", "COL")])
         y_axes <- axes$y$left[layout$SCALE_Y[panels]]
         if (params$strip.position == "left" &&
             !inside &&