+++ R/geom-boxplot.r
@@ -229,36 +229,33 @@ GeomBoxplot <- ggproto("GeomBoxplot", Geom,
       group = data$group
     )
 
-    whiskers <- new_data_frame(c(
-      list(
-        x = c(data$x, data$x),
-        xend = c(data$x, data$x),
-        y = c(data$upper, data$lower),
-        yend = c(data$ymax, data$ymin),
-        alpha = c(NA_real_, NA_real_)
-      ),
-      common
-    ), n = 2)
+    whiskers <- data_frame0(
+      x = c(data$x, data$x),
+      xend = c(data$x, data$x),
+      y = c(data$upper, data$lower),
+      yend = c(data$ymax, data$ymin),
+      alpha = c(NA_real_, NA_real_),
+      !!!common,
+      .size = 2
+    )
     whiskers <- flip_data(whiskers, flipped_aes)
 
-    box <- new_data_frame(c(
-      list(
-        xmin = data$xmin,
-        xmax = data$xmax,
-        ymin = data$lower,
-        y = data$middle,
-        ymax = data$upper,
-        ynotchlower = ifelse(notch, data$notchlower, NA),
-        ynotchupper = ifelse(notch, data$notchupper, NA),
-        notchwidth = notchwidth,
-        alpha = data$alpha
-      ),
-      common
-    ))
+    box <- data_frame0(
+      xmin = data$xmin,
+      xmax = data$xmax,
+      ymin = data$lower,
+      y = data$middle,
+      ymax = data$upper,
+      ynotchlower = ifelse(notch, data$notchlower, NA),
+      ynotchupper = ifelse(notch, data$notchupper, NA),
+      notchwidth = notchwidth,
+      alpha = data$alpha,
+      !!!common
+    )
     box <- flip_data(box, flipped_aes)
 
     if (!is.null(data$outliers) && length(data$outliers[[1]] >= 1)) {
-      outliers <- new_data_frame(list(
+      outliers <- data_frame0(
         y = data$outliers[[1]],
         x = data$x[1],
         colour = outlier.colour %||% data$colour[1],
@@ -267,8 +264,9 @@ GeomBoxplot <- ggproto("GeomBoxplot", Geom,
         size = outlier.size %||% data$size[1],
         stroke = outlier.stroke %||% data$stroke[1],
         fill = NA,
-        alpha = outlier.alpha %||% data$alpha[1]
-      ), n = length(data$outliers[[1]]))
+        alpha = outlier.alpha %||% data$alpha[1],
+        .size = length(data$outliers[[1]])
+      )
       outliers <- flip_data(outliers, flipped_aes)
 
       outliers_grob <- GeomPoint$draw_panel(outliers, panel_params, coord)