+++ R/geom-segment.r
@@ -135,7 +135,7 @@ GeomSegment <- ggproto("GeomSegment", Geom,
     starts <- subset(data, select = c(-xend, -yend))
     ends <- rename(subset(data, select = c(-x, -y)), c("xend" = "x", "yend" = "y"))
 
-    pieces <- rbind(starts, ends)
+    pieces <- vec_rbind(starts, ends)
     pieces <- pieces[order(pieces$group),]
 
     GeomPath$draw_panel(pieces, panel_params, coord, arrow = arrow,