+++ R/stat-density.r
@@ -113,25 +113,27 @@ compute_density <- function(x, w, from, to, bw = "nrd0", adjust = 1,
   # if less than 2 points return data frame of NAs and a warning
   if (nx < 2) {
     cli::cli_warn("Groups with fewer than two data points have been dropped.")
-    return(new_data_frame(list(
+    return(data_frame0(
       x = NA_real_,
       density = NA_real_,
       scaled = NA_real_,
       ndensity = NA_real_,
       count = NA_real_,
-      n = NA_integer_
-    ), n = 1))
+      n = NA_integer_,
+      .size = 1
+    ))
   }
 
   dens <- stats::density(x, weights = w, bw = bw, adjust = adjust,
     kernel = kernel, n = n, from = from, to = to)
 
-  new_data_frame(list(
+  data_frame0(
     x = dens$x,
     density = dens$y,
     scaled =  dens$y / max(dens$y, na.rm = TRUE),
     ndensity = dens$y / max(dens$y, na.rm = TRUE),
     count =   dens$y * nx,
-    n = nx
-  ), n = length(dens$x))
+    n = nx,
+    .size = length(dens$x)
+  )
 }